ugh, one of THOSE commits
you know, the ones that you make when nobody is watching?

Anyway, stuff happens when the player hits a spike pit, health is working properly, added a health pickup from the 2d game kit, and ive started (though im not sure if ill continue) a game events system, like for when the player dies. probably not necessary for this small project.