Created boss encounter
Boss has a super simple behaviour pattern in which it appears either at the left or right end of the boss room, then moves to the other end. All the behavious is in one script, though it is abstract and the state machine can easily be re-purposed elsewhere.

Created "encounter" class which tracks the state of the slime boss, locking doors, playing music, etc. easy peasy!

Create music manager to go with this ^